# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ipmoney API
  version: 0.1.0
  license:
    name: Proprietary
    url: https://ipmoney.com
  description: |
    Ipmoney 专利交易平台（小程序 + 用户 H5 + 后台）对外 API（P0）。

    业务约束（P0）：
    - 标的范围：仅 CN 专利（数据以用户上传/后台录入为准）
    - 资金路径：订金与尾款均线上支付并托管
    - 合同签署：线下签署完成后由客服/运营确认里程碑解锁尾款支付
    - 佣金：卖家承担；放款时从卖家应收中扣除
servers:
  - url: https://api.ipmoney.com/v1
...
security:
  - BearerAuth: []
paths:
...
  /public/demands/{demandId}:
    get:
      tags: [Demands]
      summary: 获取需求详情（公开）
      operationId: getPublicDemandById
      security: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemandPublic"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /search/demands:
    get:
      tags: [Search]
      summary: 产学研需求检索
      operationId: searchDemands
      security: []
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/RegionCode"
        - $ref: "#/components/parameters/IndustryTags"
        - $ref: "#/components/parameters/CooperationModes"
        - $ref: "#/components/parameters/BudgetType"
        - $ref: "#/components/parameters/BudgetMin"
        - $ref: "#/components/parameters/BudgetMax"
        - $ref: "#/components/parameters/ContentSortBy"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedDemandSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /demands:
    post:
      tags: [Demands]
      summary: 创建需求草稿（发布方）
      operationId: createDemand
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Demands]
      summary: 我的需求列表（发布方）
      operationId: listMyDemands
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ContentStatus"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedDemand"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /demands/{demandId}:
    get:
      tags: [Demands]
      summary: 获取需求详情（发布方/已登录）
      operationId: getDemandById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Demands]
      summary: 更新需求草稿（发布方）
      operationId: updateDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /demands/{demandId}/submit:
    post:
      tags: [Demands]
      summary: 提交审核（发布方）
      operationId: submitDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /demands/{demandId}/off-shelf:
    post:
      tags: [Demands]
      summary: 下架（发布方）
      operationId: offShelfDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/demands:
    get:
      tags: [Admin]
      summary: 需求审核/平台内容列表
      operationId: adminListDemandsForAudit
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/RegionCode"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/ContentStatus"
        - $ref: "#/components/parameters/ContentSource"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedDemand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: 创建平台自有需求（P1）
      operationId: adminCreateDemand
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDemandCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/demands/{demandId}:
    get:
      tags: [Admin]
      summary: 获取需求详情（后台，P1）
      operationId: adminGetDemandById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: 更新平台自有需求（后台，P1）
      operationId: adminUpdateDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDemandUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/demands/{demandId}/publish:
    post:
      tags: [Admin]
      summary: 发布需求（后台，P1）
      operationId: adminPublishDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/demands/{demandId}/off-shelf:
    post:
      tags: [Admin]
      summary: 下架需求（后台，P1）
      operationId: adminOffShelfDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/demands/{demandId}/approve:
    post:
      tags: [Admin]
      summary: 需求审核通过
      operationId: adminApproveDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/demands/{demandId}/reject:
    post:
      tags: [Admin]
      summary: 需求审核驳回
      operationId: adminRejectDemand
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DemandId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /admin/user-verifications:
    get:
      tags: [Admin]
      summary: 认证审核列表
      operationId: adminListUserVerifications
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: 按主体名称/手机号模糊搜索
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/VerificationType"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/VerificationStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedUserVerification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/user-verifications/{verificationId}/approve:
    post:
      tags: [Admin]
      summary: 认证通过
      operationId: adminApproveUserVerification
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/VerificationId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserVerification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/user-verifications/{verificationId}/reject:
    post:
      tags: [Admin]
      summary: 认证驳回
      operationId: adminRejectUserVerification
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/VerificationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserVerification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: 关键词（标题/摘要/权利人/发明人/机构名称等）
...
    BudgetType:
      name: budgetType
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/PriceType"
    BudgetMin:
      name: budgetMinFen
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MoneyFen"
    BudgetMax:
      name: budgetMaxFen
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MoneyFen"
...
    RegionCode:
      name: regionCode
      in: query
      required: false
      schema:
        type: string
        pattern: "^[0-9]{6}$"
        example: "110000"
      description: 行政区划 adcode（6 位字符串）
...
    IndustryTags:
      name: industryTags
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      description: 产业标签过滤（可多选）
...
    ContentSortBy:
      name: sortBy
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ContentSortBy"
...
    CooperationModes:
      name: cooperationModes
      in: query
      required: false
      schema:
        type: array
        items:
          $ref: "#/components/schemas/CooperationMode"
      style: form
      explode: true
      description: 合作方式过滤（可多选）
...
    ContentStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ContentStatus"
...
    ContentSource:
      name: source
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ContentSource"
      description: 内容来源过滤
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    AuditStatus:
      name: auditStatus
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/AuditStatus"
...
    DemandId:
      name: demandId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    VerificationId:
      name: verificationId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
  schemas:
    Uuid:
      type: string
      format: uuid
...
    VerificationType:
      type: string
      enum: [PERSON, COMPANY, ACADEMY, GOVERNMENT, ASSOCIATION, TECH_MANAGER]
...
    VerificationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    UserVerification:
      type: object
      required: [id, userId, type, status, submittedAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        userId:
          $ref: "#/components/schemas/Uuid"
        type:
          $ref: "#/components/schemas/VerificationType"
        status:
          $ref: "#/components/schemas/VerificationStatus"
        displayName:
          type: string
        unifiedSocialCreditCodeMasked:
          type: string
          description: 企业/机构统一社会信用代码（脱敏）
        contactName:
          type: string
        contactPhoneMasked:
          type: string
        idNumberMasked:
          type: string
          description: 个人身份证号/技术经理人证件号（脱敏）
        regionCode:
          type: string
          description: 主体所在地区（用于展示/地域推荐）
        intro:
          type: string
          description: 主体简介（用于机构展示）
        serviceTags:
          type: array
          items:
            type: string
          description: 技术经理人服务标签（仅 TECH_MANAGER）
        logoFileId:
          $ref: "#/components/schemas/Uuid"
        logoUrl:
          type: string
          format: uri
        evidenceFileIds:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
        reviewComment:
          type: string
...
    CooperationMode:
      type: string
      enum: [TRANSFER, TECH_CONSULTING, COMMISSIONED_DEV, PLATFORM_CO_BUILD]
...
    ContentMedia:
      type: object
      required: [fileId, type, sort]
      properties:
        fileId:
          $ref: "#/components/schemas/Uuid"
        type:
          type: string
          enum: [IMAGE, VIDEO, FILE]
        sort:
          type: integer
          minimum: 0
        url:
          type: string
          format: uri
          readOnly: true
        mimeType:
          type: string
          readOnly: true
        sizeBytes:
          type: integer
          format: int64
          readOnly: true
        fileName:
          type: string
          readOnly: true
    DemandSummary:
      type: object
      required: [id, title, publisher, auditStatus, status, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        source:
          $ref: "#/components/schemas/ContentSource"
        title:
          type: string
          maxLength: 200
        summary:
          type: string
          maxLength: 2000
        budgetType:
          $ref: "#/components/schemas/PriceType"
        budgetMinFen:
          $ref: "#/components/schemas/MoneyFen"
        budgetMaxFen:
          $ref: "#/components/schemas/MoneyFen"
        cooperationModes:
          type: array
          items:
            $ref: "#/components/schemas/CooperationMode"
        regionCode:
          type: string
        industryTags:
          type: array
          items:
            type: string
        keywords:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 50
        deliveryPeriod:
          $ref: "#/components/schemas/DeliveryPeriod"
        publisher:
          $ref: "#/components/schemas/OrganizationSummary"
        stats:
          $ref: "#/components/schemas/ListingStats"
        auditStatus:
          $ref: "#/components/schemas/AuditStatus"
        status:
          $ref: "#/components/schemas/ContentStatus"
        coverUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    Demand:
      allOf:
        - $ref: "#/components/schemas/DemandSummary"
        - type: object
          properties:
            publisherUserId:
              $ref: "#/components/schemas/Uuid"
            description:
              type: string
              maxLength: 2000
            keywords:
              type: array
              maxItems: 30
              items:
                type: string
                maxLength: 50
            deliveryPeriod:
              $ref: "#/components/schemas/DeliveryPeriod"
            contactName:
              type: string
              maxLength: 50
            contactTitle:
              type: string
              maxLength: 100
            contactPhoneMasked:
              type: string
              maxLength: 30
            coverFileId:
              $ref: "#/components/schemas/Uuid"
              nullable: true
            media:
              type: array
              items:
                $ref: "#/components/schemas/ContentMedia"
            aiParse:
              $ref: "#/components/schemas/AiParseResult"
              nullable: true
            updatedAt:
              type: string
              format: date-time
    DemandPublic:
      allOf:
        - $ref: "#/components/schemas/DemandSummary"
        - type: object
          properties:
            description:
              type: string
              maxLength: 2000
            media:
              type: array
              items:
                $ref: "#/components/schemas/ContentMedia"
            aiParse:
              $ref: "#/components/schemas/AiParseResult"
              nullable: true
    PagedDemandSummary:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DemandSummary"
        page:
          $ref: "#/components/schemas/PageMeta"
    PagedDemand:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Demand"
        page:
          $ref: "#/components/schemas/PageMeta"
    DemandCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        summary:
          type: string
          maxLength: 2000
        keywords:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 50
        deliveryPeriod:
          $ref: "#/components/schemas/DeliveryPeriod"
        budgetType:
          $ref: "#/components/schemas/PriceType"
        budgetMinFen:
          $ref: "#/components/schemas/MoneyFen"
        budgetMaxFen:
          $ref: "#/components/schemas/MoneyFen"
        cooperationModes:
          type: array
          items:
            $ref: "#/components/schemas/CooperationMode"
        contactName:
          type: string
          maxLength: 50
        contactTitle:
          type: string
          maxLength: 100
        contactPhoneMasked:
          type: string
          maxLength: 30
        regionCode:
          type: string
        industryTags:
          type: array
          items:
            type: string
        coverFileId:
          $ref: "#/components/schemas/Uuid"
          nullable: true
        media:
          type: array
          items:
            $ref: "#/components/schemas/ContentMedia"
    DemandUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        summary:
          type: string
          maxLength: 2000
        keywords:
          type: array
          maxItems: 30
          items:
            type: string
            maxLength: 50
        deliveryPeriod:
          $ref: "#/components/schemas/DeliveryPeriod"
        budgetType:
          $ref: "#/components/schemas/PriceType"
        budgetMinFen:
          $ref: "#/components/schemas/MoneyFen"
        budgetMaxFen:
          $ref: "#/components/schemas/MoneyFen"
        cooperationModes:
          type: array
          items:
            $ref: "#/components/schemas/CooperationMode"
        contactName:
          type: string
          maxLength: 50
        contactTitle:
          type: string
          maxLength: 100
        contactPhoneMasked:
          type: string
          maxLength: 30
        regionCode:
          type: string
        industryTags:
          type: array
          items:
            type: string
        coverFileId:
          $ref: "#/components/schemas/Uuid"
          nullable: true
        media:
          type: array
          items:
            $ref: "#/components/schemas/ContentMedia"
...
    PagedUserVerification:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserVerification"
        page:
          $ref: "#/components/schemas/PageMeta"
...
    AiParseResult:
      type: object
      required: [id, contentType, contentId, status, confidence, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        contentType:
          $ref: "#/components/schemas/AiContentType"
        contentId:
          $ref: "#/components/schemas/Uuid"
        summaryPlain:
          type: string
        featuresPlain:
          type: string
        keywords:
          type: array
          items:
            type: string
        confidence:
          type: number
          format: double
        modelVersion:
          type: string
        status:
          $ref: "#/components/schemas/AiParseStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminDemandCreateRequest:
      allOf:
        - $ref: "#/components/schemas/DemandCreateRequest"
        - type: object
          properties:
            source:
              $ref: "#/components/schemas/ContentSource"
            publisherUserId:
              $ref: "#/components/schemas/Uuid"
            auditStatus:
              $ref: "#/components/schemas/AuditStatus"
            status:
              $ref: "#/components/schemas/ContentStatus"
    AdminDemandUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/DemandUpdateRequest"
        - type: object
          properties:
            source:
              $ref: "#/components/schemas/ContentSource"
            publisherUserId:
              $ref: "#/components/schemas/Uuid"
            auditStatus:
              $ref: "#/components/schemas/AuditStatus"
            status:
              $ref: "#/components/schemas/ContentStatus"
...